refactor(code-structure): Reorganise the folder structure by separating the tests files
