+++ lib/index.ts
@@ -1,17 +1,40 @@
 // import * as path from "https://deno.land/std@0.171.0/path/mod.ts";
 
-// import { BLOG_DIR, VAULT_DIR } from "./config.ts";
+import { BLOG_DIR, VAULT_DIR } from "./config.ts";
+
+import { LinkManager } from "./linkManager.ts";
+import { Note } from "./note.ts";
+import { findFilesRecursively } from "./utils.ts";
 
 // import { Note } from "./types.ts";
 
 // await main();
 
-// async function main() {
-//   confirm(`Using the following directories:
-//     - Blog: ${BLOG_DIR},
-//     - Vault: ${VAULT_DIR},
-//     Do you want to proceed?
-// `);
+async function _main() {
+  confirm(`Using the following directories:
+    - Blog: ${BLOG_DIR},
+    - Vault: ${VAULT_DIR},
+    Do you want to proceed?
+`);
+  const notes = [];
+  const linkManager = new LinkManager();
+  const noteFilePaths = await findFilesRecursively(VAULT_DIR, {
+    match: new RegExp(".*.md$"),
+  });
+
+  const onNoteCreated = [linkManager.registerNote];
+  const handlers = [linkManager.replaceWikiLinks];
+
+  // Create all notes
+  for (const filePath of noteFilePaths) {
+    notes.push(new Note(filePath, handlers, onNoteCreated));
+  }
+
+  // Process notes
+  for (const note of notes) {
+    note.process();
+  }
+}
 
 //   try {
 //     Deno.removeSync(path.join(BLOG_DIR), { recursive: true });
